combined SCSS rendering.
now the docs and the site use the same library defined in util.js
