+++ app/templates/scripts/docs/docs.js
@@ -3,9 +3,7 @@ const tsRender = require('./ts.js');
 const htmlRender = require('./html.js');
 
 module.exports = {
-  scss: () => {
-    scssRender('src/docs/scss/docs.scss', 'dist/docs/docs.css', ['> 5% in US', 'ie >= 9']);
-  },
+  scss: scssRender,
   ts: tsRender,
   html: htmlRender
 }
